server2/requirements.txt
-i https://pypi.org/simple

cryptography==42.0.7; python_version >= '3.7'
django==5.0.6; python_version >= '3.10'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt[crypto]==5.3.1; python_version >= '3.8'
